requirements.txt
Django==1.6.2
Markdown==2.4
South==0.8.4
argparse==1.2.1
beautifulsoup4==4.3.2
bleach==1.4
boto==2.27.0
dj-database-url==0.2.2
dj-static==0.0.5
django-filter==0.7
django-storages==1.1.8
django-toolbelt==0.0.1
djangorestframework==2.3.14
gunicorn==18.0
lxml==3.2.3
psycopg2==2.5.1
requests==2.0.0
static==0.4
wsgiref==0.1.2
